updated project structure
