+++ scripts/Pipfile
@@ -8,6 +8,7 @@ et-xmlfile = "*"
 openpyxl = "*"
 pyyaml = "*"
 parsy = "*"
+ci = {editable = true, path = "."}
 
 [dev-packages]
