+++ scripts/pyproject.toml
@@ -1,0 +1,7 @@
+# build configs
+[build-system]
+requires = [
+    "setuptools>=42",
+    "wheel"
+]
+build-backend = "setuptools.build_meta"
