+++ scripts/requirements.txt
@@ -1,4 +1,6 @@
+-e .
 et-xmlfile==1.1.0
 openpyxl==3.0.9
+packaging==21.0
 parsy==1.3.0
 PyYAML==5.4.1
