+++ scripts/setup.cfg
@@ -1,0 +1,25 @@
+[metadata]
+name = ci
+version = 0.0.1
+author = @S7rizh, @nikololiahim
+author_email = olomishcak@outlook.com
+description = A package containing CI jobs to run in the AwesomeBugs repository
+long_description = file: README.md
+license_files = LICENSE.md
+long_description_content_type = text/markdown
+url = https://github.com/S7rizh/AwesomeBugs
+project_urls =
+    Bug Tracker = https://github.com/S7rizh/AwesomeBugs/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.9
+
+[options.packages.find]
+where = src
