+++ scripts/setup.py
@@ -1,0 +1,5 @@
+import setuptools
+
+setuptools.setup(
+    zip_safe=False
+)
