+++ scripts/src/analyze_reports/__init__.py
@@ -1,0 +1,1 @@
+from scripts.src.analyze_reports.report import AnalyzerReport, AnalyzerReportRow
