+++ scripts/src/clang_tidy/__main__.py
@@ -229,7 +229,7 @@ def main():
                         help='Run clang-tidy in quiet mode')
     args = parser.parse_args()
 
-    db_path = '../build/compile_commands.json'
+    db_path = 'compile_commands.json'
 
     if args.build_path is not None:
         build_path = args.build_path
