+++ scripts/src/report_parsers/__init__.py
@@ -1,0 +1,3 @@
+from scripts.src.report_parsers.clang_tidy_parser import ClangTidyParser
+from scripts.src.report_parsers.coverity_parser import CoverityParser
+
